Modificada vista de pagina, agregado pagina about me, modificando urls y templates correspondientes

+++ blog/views.py
@@ -1,29 +1,42 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse
 from blog.models import Articulo
 from blog.forms import ArticuloForm
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.http import HttpResponse
-
+from django.contrib import messages
+from django.contrib.auth.models import User
 
 # Create your views here.
+def about(request):
+    contexto = {
+        'nombre': 'Gabriela',
+        'edad': 42,
+        'pais': 'Argentina',
+        'profesion': 'Ingeniera en Telecomunicaciones',
+        'titulo': 'Sobre mí',
+        'descripcion': 'Soy secretaria en una escuela primaria, me gustan las manualidades, empecé el curso de Python por una amiga que me insistió, me gusta pero tengo que dedicar mucho tiempo y a veces con mis hijas y trabajo, no puedo.',
+    }
+
+    return render(
+        request=request,
+        template_name= 'blog/about.html',
+        context=contexto,
+    )
+
 
 @login_required
 def crear_articulo(request):
     if request.method == 'POST':
-        form = ArticuloForm(request.POST)
+        form = ArticuloForm(request.POST, request.FILES)
         if form.is_valid():
-            data= form.cleaned_data
-            titulo= data["titulo"]
-            subtitulo= data["subtitulo"]
-            cuerpo= data["cuerpo"]
-            autor= request.user
-            articulo= Articulo.objects.create(titulo=titulo, subtitulo= subtitulo, cuerpo= cuerpo, autor=autor)
+            articulo= form.save()
             url_exitosa=reverse('lista_articulos') #redirecciono a la lista de articulos
             return redirect(url_exitosa)
     else:
         form = ArticuloForm()
+
     http_response= render(
         request= request, 
         template_name= 'blog/crear_articulo.html', 
@@ -80,29 +93,37 @@ def ver_articulo(request, id):
 
 
 @login_required
-def editar_articulo(request, id): 
-    articulo= Articulo.objects.get(id= id)
+def editar_articulo(request, id_articulo):
+    articulo = get_object_or_404(Articulo, id=id_articulo)
 
-    if request.method=="POST":
-        form= ArticuloForm(request.POST, instance=articulo)
+    if request.user != articulo.autor:
+        messages.error(request, "No tienes permiso para editar este artículo.")
+        return redirect('inicio')
 
+    if request.method == "POST":
+        form = ArticuloForm(request.POST, instance=articulo)
         if form.is_valid():
-            articulo.save()
-            return redirect('lista_articulos') 
+            form.save()
+            messages.success(request, "El artículo ha sido actualizado.")
+            return redirect('inicio')
     else:
-        form= ArticuloForm(instance=articulo)
-        
+        form = ArticuloForm(instance=articulo)
+
     return render(
         request=request,
-        template_name= 'blog/crear_articulo.html',
-        context={'form': form},
+        template_name='blog/editar_articulo.html',
+        context={'form': form, 'articulo': articulo},
     )
-
 
 
 @login_required
 def eliminar_articulo(request, id): #pasamos id como argumento para saber que articulo es
     articulo= Articulo.objects.get(id= id)
+
+    if articulo.autor != request.user:  # Verificar si el usuario no es el autor del artículo
+        messages.add_message(request, messages.ERROR, "No puedes eliminar este artículo porque no es tuyo.", extra_tags=str(articulo.id))
+        return redirect('lista_articulos')
+    
     if request.method=="POST":
         articulo.delete() #borra el articulo
         url_exitosa=reverse('lista_articulos') #manda a url_exitosa donde ya no esta ese articulo
